requirements.txt
asgiref==3.6.0
assertpy==1.1
attrs==23.1.0
beautifulsoup4==4.11.1
black==22.10.0
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.6
Django==4.2.4
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-rest-framework==0.1.0
django-tqdm==1.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.1
flake8==5.0.4
Flask==1.1.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.4.1
MarkupSafe==1.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
outcome==1.2.0
packaging==23.1
pathspec==0.10.1
platformdirs==2.5.3
pluggy==1.2.0
psycopg2==2.9.7
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.0
pytz==2022.7.1
requests==2.28.1
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.19
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
trio==0.22.2
trio-websocket==0.10.3
typing-extensions==4.4.0
tzdata==2023.3
urllib3==1.26.12
Werkzeug==1.0.1
wsproto==1.2.0
zipp==3.11.0
